[test][IRGen] Split out objc-interop specific part in abitypes.swift
