+++ test/IRGen/Inputs/abi/module.modulemap
@@ -1,2 +1,3 @@
 module gadget { header "Gadget.h" }
+module c_gadget { header "CGadget.h" }
 module c_layout { header "c_layout.h" }
